+++ src/db_migrate.cpp
@@ -13,7 +13,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.2.13"};
+constexpr auto version = std::string_view{"0.2.14"};
 
 auto print_help() -> void
 {